Move lib/srv/uacc to session/uacc
